Fix Build Fail and Modify Navbar
